+++ R/control.R
@@ -139,8 +139,8 @@ control.dcm <- function(type, nsteps, dt = 1, odemethod = "rk4",
 #'        function of \code{\link{departures.icm}}.
 #' @param arrivals.FUN Module to simulate arrivals or entries, with the default
 #'        function of \code{\link{arrivals.icm}}.
-#' @param get_prev.FUN Module to calculate disease prevalence at each time step,
-#'        with the default function of \code{\link{get_prev.icm}}.
+#' @param prevalence.FUN Module to calculate disease prevalence at each time step,
+#'        with the default function of \code{\link{prevalence.icm}}.
 #' @param verbose If \code{TRUE}, print model progress to the console.
 #' @param verbose.int Time step interval for printing progress to console, where
 #'        0 (the default) prints completion status of entire simulation and
@@ -186,7 +186,7 @@ control.icm <- function(type, nsteps, nsims = 1, rec.rand = TRUE, a.rand = TRUE,
                         d.rand = TRUE, initialize.FUN = initialize.icm,
                         infection.FUN = NULL, recovery.FUN = NULL,
                         departures.FUN = NULL, arrivals.FUN = NULL,
-                        get_prev.FUN = NULL, verbose = FALSE,
+                        prevalence.FUN = NULL, verbose = FALSE,
                         verbose.int = 0, skip.check = FALSE, ...) {
 
   # Get arguments
@@ -278,13 +278,6 @@ control.icm <- function(type, nsteps, nsims = 1, rec.rand = TRUE, a.rand = TRUE,
 #'        distribution with the probability equal to the governing departure rates.
 #'        If \code{FALSE}, then a deterministic rounded count of the expectation
 #'        implied by those rates.
-#' @param tea.status If \code{TRUE}, use a temporally extended attribute (TEA)
-#'        to store disease status. A TEA is needed for plotting static networks
-#'        at different time steps and for animating dynamic networks with evolving
-#'        status. TEAs are computationally inefficient for large simulations and
-#'        should be toggled off in those cases. This argument automatically set
-#'        to \code{FALSE} if \code{tgl=TRUE}.
-#'
 #' @param tgl Logical indicating usage of either \code{tergm} (\code{tgl = TRUE}),
 #'        or \code{tergmLite} (\code{tgl = FALSE}). Default of \code{FALSE}.
 #' @param attr.rules A list containing the  rules for setting the attributes of
@@ -294,13 +287,6 @@ control.icm <- function(type, nsteps, nsims = 1, rec.rand = TRUE, a.rand = TRUE,
 #'        which subgroup epidemic prevalences should be calculated. This nodal
 #'        attribute must be contained in the network model formation formula,
 #'        otherwise it is ignored.
-#' @param use.pids If \code{TRUE}, use persistent ids for vertices; otherwise,
-#'        numeric ids will be recycled in models with vital dynamics. For one-group
-#'        simulations, this will be a random hexidecimal value; for two-group
-#'        simulations, it will be based on \code{pid.prefix}.
-#' @param pid.prefix For two-group network simulations with vital dynamics,
-#'        a character vector of length 2 containing the prefixes, with the
-#'        default of \code{c("F", "M")}.
 #' @param initialize.FUN Module to initialize the model at time 1, with the
 #'        default function of \code{\link{initialize.net}}.
 #' @param departures.FUN Module to simulate departure or exit, with the default function
@@ -313,8 +299,11 @@ control.icm <- function(type, nsteps, nsims = 1, rec.rand = TRUE, a.rand = TRUE,
 #'        with the default function of \code{\link{resim_nets}}.
 #' @param infection.FUN Module to simulate disease infection, with the default
 #'        function of \code{\link{infection.net}}.
-#' @param get_prev.FUN Module to calculate disease prevalence at each time step,
-#'        with the default function of \code{\link{get_prev.net}}.
+#' @param nwupdate.FUN Module to handle updating of network structure and nodal
+#'        attributes due to exogenous epidemic model processes, with the default
+#'        function of \code{\link{nwupdate.net}}.
+#' @param prevalence.FUN Module to calculate disease prevalence at each time step,
+#'        with the default function of \code{\link{prevalence.net}}.
 #' @param verbose.FUN Module to print simulation progress to screen, with the
 #'        default function of \code{\link{verbose.net}}.
 #' @param module.order A character vector of module names that lists modules the
@@ -353,6 +342,8 @@ control.icm <- function(type, nsteps, nsims = 1, rec.rand = TRUE, a.rand = TRUE,
 #'        and control settings before running base epidemic models. Setting
 #'        this to \code{FALSE} is recommended when running models with new modules
 #'        specified.
+#' @param raw_output If \code{TRUE}, \code{netsim} will output a list of nestsim
+#'        Data (one per simulation) instead of a formatted \code{netsim} object.
 #' @param ... Additional control settings passed to model.
 #'
 #' @details
@@ -410,30 +401,40 @@ control.icm <- function(type, nsteps, nsims = 1, rec.rand = TRUE, a.rand = TRUE,
 #'
 #' @export
 #'
-
 control.net <- function(type,
-                        nsteps, start = 1,
-                        nsims = 1, ncores = 1,
-                        depend, rec.rand = TRUE,
-                        a.rand = TRUE, d.rand = TRUE,
-                        tea.status = TRUE, tgl = FALSE,
-                        attr.rules, epi.by,
-                        use.pids = TRUE, pid.prefix,
+                        nsteps,
+                        start = 1,
+                        nsims = 1,
+                        ncores = 1,
+                        depend,
+                        rec.rand = TRUE,
+                        a.rand = TRUE,
+                        d.rand = TRUE,
+                        tgl = FALSE,
+                        attr.rules,
+                        epi.by,
                         initialize.FUN = initialize.net,
                         resim_nets.FUN = resim_nets,
                         infection.FUN = NULL,
                         recovery.FUN = NULL,
                         departures.FUN = NULL,
                         arrivals.FUN = NULL,
-                        nw.update.FUN = nw.update.net,
-                        get_prev.FUN = NULL,
+                        nwupdate.FUN = nwupdate.net,
+                        prevalence.FUN = NULL,
                         verbose.FUN = verbose.net,
                         module.order = NULL,
-                        set.control.ergm, set.control.stergm,
-                        save.nwstats = TRUE, nwstats.formula = "formation",
-                        save.transmat = TRUE, save.network = TRUE,
-                        save.other, verbose = TRUE,
-                        verbose.int = 1, skip.check = FALSE, ...) {
+                        set.control.ergm,
+                        set.control.stergm,
+                        save.nwstats = TRUE,
+                        nwstats.formula = "formation",
+                        save.transmat = TRUE,
+                        save.network = TRUE,
+                        save.other,
+                        verbose = TRUE,
+                        verbose.int = 1,
+                        skip.check = FALSE,
+                        raw_output = FALSE,
+                        ...) {
 
   # Get arguments
   p <- list()
@@ -478,11 +479,20 @@ control.net <- function(type,
     }
   }
 
+  # Using tergmLite --> depend = TRUE
+  if (tgl == TRUE) {
+    p$depend <- TRUE
+  }
+
+  # Temporary until we develop a nwstats fix for tergmLite
+  if (tgl == TRUE) {
+    p$save.nwstats <- FALSE
+  }
+
   ## Defaults and checks
 
   #Check whether any base modules have been redefined by user (note: must come after above)
-  bi.nms <- bi.nms[-which(bi.nms %in% c("initialize.FUN", "edges_correct.FUN",
-                                        "resim_nets.FUN", "verbose.FUN", "nw.update.FUN"))]
+  bi.nms <- bi.nms[-which(bi.nms %in% c("initialize.FUN", "resim_nets.FUN", "verbose.FUN", "nwupdate.FUN"))]
   if (length(bi.nms) > 0){
     flag1 <- logical()
     for (args in 1:length(bi.nms)) {
@@ -525,10 +535,6 @@ control.net <- function(type,
     p$set.control.ergm <- control.simulate.ergm(MCMC.burnin = 2e5)
   }
 
-  if (p$tgl == TRUE) {
-    p$tea.status <- FALSE
-  }
-
   if (is.null(p$type)) {
     names <- unlist(lapply(sys.call()[-1], as.character))
     pos <- which(names(names) %in% grep(".FUN", names(names), value = TRUE))
@@ -537,15 +543,15 @@ control.net <- function(type,
   }
 
   if (p$type != "SIR" && !is.null(p$type)) {
-    p$f.names <- c("arrivals.FUN", "departures.FUN", "infection.FUN", "get_prev.FUN")
-    p$f.args  <- c("arrivals.net", "departures.net", "infection.net", "get_prev.net")
+    p$f.names <- c("arrivals.FUN", "departures.FUN", "infection.FUN", "prevalence.FUN")
+    p$f.args  <- c("arrivals.net", "departures.net", "infection.net", "prevalence.net")
   }
 
   if (p$type %in% c("SIR", "SIS") && !is.null(p$type)) {
     p$f.names <- c("arrivals.FUN", "departures.FUN", "infection.FUN",
-                   "recovery.FUN", "get_prev.FUN")
+                   "recovery.FUN", "prevalence.FUN")
     p$f.args  <- c("arrivals.net", "departures.net", "infection.net",
-                   "recovery.net", "get_prev.net")
+                   "recovery.net", "prevalence.net")
   }
 
   ## Output