+++ tests/testthat/test-merge.R
@@ -98,14 +98,13 @@ test_that("merge for netsim", {
 })
 
 test_that("merge works for open sims saving nw stats", {
-  skip_on_cran()
-  nw <- network.initialize(n = 50, directed = FALSE)
-  est <- netest(nw, formation = ~edges, target.stats = 10,
-                coef.diss = dissolution_coefs(~offset(edges), 10, 0),
+  nw <- network.initialize(n = 100, directed = FALSE)
+  est <- netest(nw, formation = ~edges, target.stats = 20,
+                coef.diss = dissolution_coefs(~offset(edges), 10, 0.01),
                 verbose = FALSE)
   param <- param.net(inf.prob = 0.9, a.rate = 0.01, ds.rate = 0.01, di.rate = 0.01)
   init <- init.net(i.num = 1)
-  control <- control.net(type = "SI", nsteps = 5, save.stats = TRUE,
+  control <- control.net(type = "SI", nsteps = 5, save.nwstats = TRUE,
                          nwstats.formula = ~edges + meandeg + degree(0) + concurrent,
                          verbose = FALSE)
 
@@ -117,6 +116,3 @@ test_that("merge works for open sims saving nw stats", {
   expect_true(all(sapply(z$stats$nwstats, dim)[1,] == 5) &
               all(sapply(z$stats$nwstats, dim)[2,] == 4))
 })
-
-
-